+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.0.2
+
+- Cache should invalidate when 
+- Setup cache invalidation to check if users groq query has changed, if so cache is invalid and the loader will fetch
+- Setup tests that don't interfere with the manual tests (tests that relied on certain number of posts to be created)
+
 ## 1.0.1 
 
 Change logging so it outputs logs for whether or not content was loaded from cache or fetched.